Chapter 4. Exercise 8.
You are given a binary tree in which each node contains a value. Design
an algorithm to print all paths which sum up to that value. Note that it
can be any path in the tree - it does not have to start at the root.

Comments:

I did it with two recursions, so it will be very memory
expensive. But the code is pretty clean and it is working.